+++ packages/core/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @redocly/openapi-core
 
+## 2.28.1
+
 ## 2.28.0
 
 ### Minor Changes